Created an API and a reload command
Created an API called "SkinAPI" for other plugin developers and a reload
command with a permission node. Also changed updater method and added
the real project ID. Lastly, made sure any commands throwing errors will
not cause the server to crash and only prints it to the console, telling
the player there was an error.