Fixed missing import and the SkinAPI
Fixed the missing import of the "Lang" class in "CommandBase" and also
changed SkinAPI so that it does not override the current features when
using the set() methods. Also listened to PluginEnableEvent so that the
plugin can be enabled if ProtocolLib is loaded. Lastly, made it so that
when the server reloads and the plugin is enabled again, set the join
skins if the player has permission to.